requirements.txt
numpy
pandas
matplotlib
seaborn
scikit-learn
scipy
python-docx
joblib
